Improve lldb-dap connection
